Enhance telemetry and observability in order management system
- Introduced OpenTelemetry for tracing and metrics collection, including custom spans and metrics for order creation.
- Updated the Order model to emit an OrderCreated event with relevant metrics.
- Implemented OrderMetricsSubscriber to record metrics upon order creation.
- Enhanced CreateOrderHandler to include telemetry spans for better observability during order processing.
- Updated README.md to document the new shared telemetry layer and observability features.
- Refactored service configuration to include telemetry interfaces and dependencies.

+++ api/src/Domain/Model/OrderItem.php
@@ -41,6 +41,6 @@ final readonly class OrderItem
 
     public function totalCents(): int
     {
-        return (int) ($this->quantity * $this->unitPriceCents);
+        return $this->quantity * $this->unitPriceCents;
     }
 }
